Move config.ts and types.ts out of common/
