+++ src/dns-logic/diff.ts
@@ -2,7 +2,7 @@ import { SetUtil } from "../deps.ts";
 
 import type {
   BaseRecord, ZoneState, DnsProvider, RecordGroupDiff,
-} from "../common/types.ts";
+} from "../types.ts";
 
 // TODO: the grouping/comparision keys should probably be enriched once directly onto the records
 export function buildDiff<Trecord extends BaseRecord>(state: ZoneState<Trecord>, rules: Pick<DnsProvider<Trecord>, 'GroupingKey' | 'ComparisionKey'>): Array<RecordGroupDiff<Trecord>> {
