+++ src/providers/vultr/mod_test.ts
@@ -1,6 +1,6 @@
 import { assertEquals } from "https://deno.land/std@0.115.0/testing/asserts.ts";
 
-import { SourceRecord, ZoneState } from "../../common/types.ts";
+import { SourceRecord, ZoneState } from "../../types.ts";
 import { buildDiff } from "../../dns-logic/diff.ts";
 import { VultrApiMock } from "./mock.ts";
 import { VultrProvider, VultrRecord } from "./mod.ts";
